refactor(dbal): rename all DBAL classes
