+++ classes/PunchCMS/DBAL/ContentLanguage.php
@@ -1,69 +1,68 @@
 <?php
 
-/***
- *
- * Template DBA Class.
+namespace PunchCMS\DBAL;
+
+/**
+ * ContentLanguage DBA Class.
+ * @author Felix Langfeldt <felix@neverwoods.com>
+ * @internal
  *
  */
-
-class DBA_Template extends DBA__Object {
+class ContentLanguage extends Object
+{
 	protected $id = NULL;
 	protected $accountid = 0;
 	protected $name = "";
-	protected $apiname = "";
-	protected $description = "";
-	protected $parentid = 0;
-	protected $ispage = 0;
-	protected $iscontainer = 0;
-	protected $forcecreation = 0;
+	protected $abbr = "";
 	protected $active = 0;
+	protected $default = 0;
 	protected $username = "";
 
 	//*** Constructor.
-	public function DBA_Template() {
-		self::$object = "Template";
-		self::$table = "pcms_template";
+	public function DBA_ContentLanguage() {
+		self::$object = "ContentLanguage";
+		self::$table = "pcms_language";
 	}
 
 	//*** Static inherited functions.
 	public static function selectByPK($varValue, $arrFields = array(), $accountId = NULL) {
-		self::$object = "Template";
-		self::$table = "pcms_template";
+		self::$object = "ContentLanguage";
+		self::$table = "pcms_language";
 
 		return parent::selectByPK($varValue, $arrFields, $accountId);
 	}
 
 	public static function select($strSql = "") {
-		self::$object = "Template";
-		self::$table = "pcms_template";
+		self::$object = "ContentLanguage";
+		self::$table = "pcms_language";
 
 		return parent::select($strSql);
 	}
 
 	public static function doDelete($varValue) {
-		self::$object = "Template";
-		self::$table = "pcms_template";
+		self::$object = "ContentLanguage";
+		self::$table = "pcms_language";
 
 		return parent::doDelete($varValue);
 	}
 
 	public function save($blnSaveModifiedDate = TRUE) {
-		self::$object = "Template";
-		self::$table = "pcms_template";
+		self::$object = "ContentLanguage";
+		self::$table = "pcms_language";
 
 		return parent::save($blnSaveModifiedDate);
 	}
 
 	public function delete($accountId = NULL) {
-		self::$object = "Template";
-		self::$table = "pcms_template";
+		self::$object = "ContentLanguage";
+		self::$table = "pcms_language";
 
 		return parent::delete($accountId);
 	}
 
 	public function duplicate() {
-		self::$object = "Template";
-		self::$table = "pcms_template";
+		self::$object = "ContentLanguage";
+		self::$table = "pcms_language";
 
 		return parent::duplicate();
 	}
