+++ classes/PunchCMS/DBAL/Element.php
@@ -1,65 +1,76 @@
 <?php
 
-/***
+namespace PunchCMS\DBAL;
+
+/**
  *
- * StorageItem DBA Class.
+ * Element DBA Class.
+ * @author Felix Langfeldt <felix@neverwoods.com>
+ * @internal
  *
  */
-
-class DBA_StorageItem extends DBA__Object {
+class Element extends Object
+{
 	protected $id = NULL;
 	protected $accountid = 0;
-	protected $parentid = 0;
 	protected $name = "";
+	protected $namecount = 0;
+	protected $apiname = "";
 	protected $description = "";
 	protected $typeid = 0;
+	protected $templateid = 0;
+	protected $ispage = 0;
+	protected $parentid = 0;
+	protected $userid = 0;
+	protected $groupid = 0;
+	protected $active = 1;
 	protected $username = "";
 
 	//*** Constructor.
-	public function DBA_StorageItem() {
-		self::$object = "StorageItem";
-		self::$table = "pcms_storage_item";
+	public function DBA_Element() {
+		self::$object = "Element";
+		self::$table = "pcms_element";
 	}
 
 	//*** Static inherited functions.
 	public static function selectByPK($varValue, $arrFields = array(), $accountId = NULL) {
-		self::$object = "StorageItem";
-		self::$table = "pcms_storage_item";
+		self::$object = "Element";
+		self::$table = "pcms_element";
 
 		return parent::selectByPK($varValue, $arrFields, $accountId);
 	}
 
 	public static function select($strSql = "") {
-		self::$object = "StorageItem";
-		self::$table = "pcms_storage_item";
+		self::$object = "Element";
+		self::$table = "pcms_element";
 
 		return parent::select($strSql);
 	}
 
 	public static function doDelete($varValue) {
-		self::$object = "StorageItem";
-		self::$table = "pcms_storage_item";
+		self::$object = "Element";
+		self::$table = "pcms_element";
 
 		return parent::doDelete($varValue);
 	}
 
 	public function save($blnSaveModifiedDate = TRUE) {
-		self::$object = "StorageItem";
-		self::$table = "pcms_storage_item";
+		self::$object = "Element";
+		self::$table = "pcms_element";
 
 		return parent::save($blnSaveModifiedDate);
 	}
 
 	public function delete($accountId = NULL) {
-		self::$object = "StorageItem";
-		self::$table = "pcms_storage_item";
+		self::$object = "Element";
+		self::$table = "pcms_element";
 
 		return parent::delete($accountId);
 	}
 
 	public function duplicate() {
-		self::$object = "StorageItem";
-		self::$table = "pcms_storage_item";
+		self::$object = "Element";
+		self::$table = "pcms_element";
 
 		return parent::duplicate();
 	}
