+++ classes/PunchCMS/DBAL/ElementMeta.php
@@ -1,63 +1,66 @@
 <?php
 
-/***
- *
- * ElementFieldBigText DBA Class.
+namespace PunchCMS\DBAL;
+
+/**
+ * ElementMeta DBA Class.
+ * @author Felix Langfeldt <felix@neverwoods.com>
+ * @internal
  *
  */
-
-class DBA_ElementFieldBigText extends DBA__Object {
+class ElementMeta extends Object {
 	protected $id = NULL;
+	protected $elementid = 0;
+	protected $name = "";
 	protected $value = "";
-	protected $fieldid = 0;
 	protected $languageid = 0;
 	protected $cascade = 0;
 
 	//*** Constructor.
-	public function DBA_ElementFieldBigText() {
-		self::$object = "ElementFieldBigText";
-		self::$table = "pcms_element_field_bigtext";
+	public function DBA_ElementMeta() {
+		self::$object = "ElementMeta";
+		self::$table = "pcms_element_meta";
 	}
 
 	//*** Static inherited functions.
 	public static function selectByPK($varValue, $arrFields = array(), $accountId = NULL) {
-		self::$object = "ElementFieldBigText";
-		self::$table = "pcms_element_field_bigtext";
+		self::$object = "ElementMeta";
+		self::$table = "pcms_element_meta";
 
 		return parent::selectByPK($varValue, $arrFields, $accountId);
 	}
 
 	public static function select($strSql = "") {
-		self::$object = "ElementFieldBigText";
-		self::$table = "pcms_element_field_bigtext";
+		self::$object = "ElementMeta";
+		self::$table = "pcms_element_meta";
 
 		return parent::select($strSql);
 	}
 
 	public static function doDelete($varValue) {
-		self::$object = "ElementFieldBigText";
-		self::$table = "pcms_element_field_bigtext";
+		self::$object = "ElementMeta";
+		self::$table = "pcms_element_meta";
 
 		return parent::doDelete($varValue);
 	}
 
 	public function save($blnSaveModifiedDate = TRUE) {
-		self::$object = "ElementFieldBigText";
-		self::$table = "pcms_element_field_bigtext";
+		self::$object = "ElementMeta";
+		self::$table = "pcms_element_meta";
 
 		return parent::save($blnSaveModifiedDate);
 	}
 
 	public function delete($accountId = NULL) {
-		self::$object = "ElementFieldBigText";
-		self::$table = "pcms_element_field_bigtext";
+		self::$object = "ElementMeta";
+		self::$table = "pcms_element_meta";
 
 		return parent::delete($accountId);
 	}
 
 	public function duplicate() {
-		self::$object = "ElementFieldBigText";
-		self::$table = "pcms_element_field_bigtext";
+		self::$object = "ElementMeta";
+		self::$table = "pcms_element_meta";
 
 		return parent::duplicate();
 	}
